+++ iati/codelists.py
@@ -9,6 +9,7 @@ class Codelist(object):
     """Representation of a Codelist as defined within the IATI SSOT.
 
     Attributes:
+        complete (bool): Whether the Codelist is complete or not. If complete, attributes making use of this Codelist must only contain values present on the Codelist. If not complete, this is merely strongly advised.
         codes (:obj:`set` of :obj:`iati.Code`): The codes demonstrating the range of values that the Codelist may represent.
         name (str): The name of the Codelist.
 
@@ -63,7 +64,7 @@ class Codelist(object):
             self.name = tree.attrib['name']
             for code_el in tree.findall('codelist-items/codelist-item'):
                 value = code_el.findtext('code')
-                name = code_el.findtext('name/narrative')
+                name = code_el.findtext('name/narrative') or code_el.findtext('name')
 
                 if (value is None) and (name is None):
                     msg = "The provided Codelist ({0}) has a Code that does not contain a name or value.".format(self.name)
@@ -75,6 +76,12 @@ class Codelist(object):
                     name = ''
                 self.codes.add(iati.Code(value, name))
 
+            try:
+                self.complete = True if tree.attrib['complete'] == '1' else False
+            except KeyError:
+                pass
+
+        self.complete = None
         self.codes = set()
         self.name = name
 
@@ -85,7 +92,6 @@ class Codelist(object):
         self._url = None
         self._ref = None
         self._category_codelist = None
-        self._complete = None
 
         if xml:
             parse_from_xml(xml)
@@ -175,7 +181,7 @@ class Code(object):
     """
 
     # pylint: disable=too-many-instance-attributes
-    def __init__(self, value=None, name=None):
+    def __init__(self, value=None, name=''):
         """Initialise a Code.
 
         Args:
@@ -209,8 +215,13 @@ class Code(object):
         Todo:
             Utilise all attributes as part of the equality process.
 
+            Test comparison with strings.
+
         """
-        return ((self.name) == (other.name)) and ((self.value) == (other.value))
+        try:
+            return ((self.name) == (other.name)) and ((self.value) == (other.value))
+        except AttributeError:
+            return self.value == other
 
     def __ne__(self, other):
         """Check Code inequality.
@@ -229,8 +240,10 @@ class Code(object):
         Todo:
             Utilise all attributes as part of the hashing process.
 
+            Be able to deal with checks against both Codes and strings.
+
         """
-        return hash((self.name, self.value))
+        return hash((self.value))
 
     @property
     def xsd_enumeration(self):